DS-3801-group by frankgraave: seperate eventsubscriber for redirect. cm+ now able to see and edit closed groups
