+++ modules/social_features/social_group/social_group.services.yml
@@ -10,3 +10,7 @@ services:
     class: Drupal\social_group\EventSubscriber\AddressFormatSubscriber
     tags:
       - { name: event_subscriber }
+  social_group.redirect_subscriber:
+    class: Drupal\social_group\EventSubscriber\RedirectSubscriber
+    tags:
+      - { name: event_subscriber }
